more Score Manager
I tried to connect the scoreText object to the text on the screen. I do not have the updated screens on my unity, don't know why, the next step will be to fix that. Some bugs I have are about objects not being found and a reference being required for a non-static field.